+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.1.0] - 2025-08-13
+
+### Changed
+
+- Adds the concept and use of temporary attribute metadata overrides. Temporary attributes can be
+  defined in the earthdata-varinfo configuration by using a prefix of `_*` on the attribute name.
+  The metadata-annotator makes these attributes accessible in the varinfo object to assist with
+  other annotations, but prevents these from being written to the datatree object and netcdf output
+  file.
+- Updates `master_geotransform`, `corner_point_offsets`, `subset_index_reference` from persistent
+  attribute annotations to temporary attributes.
+- Add Developer Notes section to README to detail the configuration required for creating spatial
+  dimension variables and explain temporary attributes.
+
+
 ## [v1.0.4] - 2025-07-22
 
 ### Changed
@@ -65,6 +80,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   for a variable as specified via an `earthdata-varinfo` configuration file.
 - Initial repository setup with utility scripts and Dockerfiles.
 
+[v1.0.5]: https://github.com/nasa/harmony-metadata-annotator/releases/tag/1.0.5
 [v1.0.4]: https://github.com/nasa/harmony-metadata-annotator/releases/tag/1.0.4
 [v1.0.3]: https://github.com/nasa/harmony-metadata-annotator/releases/tag/1.0.3
 [v1.0.2]: https://github.com/nasa/harmony-metadata-annotator/releases/tag/1.0.2